Task #5038: fix size_t vs unsigned confusion in Kernel objs as kernel arg. Also convert most size_t to unsigned, except for buffer sizes, #ops, and #transferred calc. Force size_t calculation on buffer size mults.

+++ trunk/RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerTransposeKernel.cc
@@ -90,8 +90,8 @@ namespace LOFAR
       case BeamFormerTransposeKernel::INPUT_DATA: 
       case BeamFormerTransposeKernel::OUTPUT_DATA:
         return
-          itsParameters.nrChannelsPerSubband * itsParameters.nrSamplesPerChannel * 
-          NR_POLARIZATIONS * itsParameters.nrTABs * sizeof(std::complex<float>);
+          (size_t) itsParameters.nrChannelsPerSubband * itsParameters.nrSamplesPerChannel * 
+            NR_POLARIZATIONS * itsParameters.nrTABs * sizeof(std::complex<float>);
       default:
         THROW(GPUProcException, "Invalid bufferType (" << bufferType << ")");
       }
